backend/requirements.txt
fastapi
pandas
python-dotenv
uvicorn
pyyaml
requests
python-multipart
